pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "breez-mcp"
version = "0.1.0"
description = "Breez MCP Server - FastMCP implementation for Lightning wallet functionality"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Breez MCP Contributors"}
]
keywords = ["mcp", "lightning", "breez", "bitcoin", "wallet"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    "fastmcp>=0.4.0",
    "breez-sdk-spark>=0.2.7",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/breez/breez-mcp"
Repository = "https://github.com/breez/breez-mcp"
Documentation = "https://github.com/breez/breez-mcp#readme"

[project.scripts]
breez-mcp = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."